chore: rename paymentIntent to intent
