+++ .changeset/smart-baboons-buy.md
@@ -2,22 +2,23 @@
 "paymongo.js": major
 ---
 
-# v2 api
+# v2.0.0
 
 ### Breaking Changes
 
-- The structure of the api for v2 has renamed most of the functions. We recommend reading the usage guide for v2.
+- The structure of the api for v2 has renamed some of the functions. We recommend reading the usage guide for v2.
 
 ### Log
 
-* 3f8e2e9 chore: add retrieve payment and update readme (princejoogie)
-* 888689a chore: add workflow secret (princejoogie)
-* 42dcdda chore: add release script (princejoogie)
-* de29a22 chore: add changeset and workflows (princejoogie)
-* 3467886 feat: payment-intent/create (princejoogie)
-* 6702145 chore: initial structure for v2 (princejoogie)
-* 2b857e2 chore: update workflow (princejoogie)
-* d0fe2a3 chore: add size-limit dependency (princejoogie)
-* 1670631 chore: update packages (princejoogie)
-* 4c0bbcb chore: remove unused deps (princejoogie)
-* 036c4d0 chore: migrate to yarn (princejoogie)
+* 767f49a chore: access public on changeset (@princejoogie)
+* e372748 fix: export type (@princejoogie)
+* e80e10f chore: update changeset (@princejoogie)
+* 3f8e2e9 chore: add retrieve payment and update readme (@princejoogie)
+* 888689a chore: add workflow secret (@princejoogie)
+* 42dcdda chore: add release script (@princejoogie)
+* de29a22 chore: add changeset and workflows (@princejoogie)
+* 3467886 feat: payment-intent/create (@princejoogie)
+* 6702145 chore: initial structure for v2 (@princejoogie)
+* 2b857e2 chore: update workflow (@princejoogie)
+* d0fe2a3 chore: add size-limit dependency (@princejoogie)
+* 1670631 chore: update packages (@princejoogie)
