+++ src/client.ts
@@ -1,6 +1,6 @@
 import { api } from "./api/base";
-import { createPaymentIntent } from "./api/payment-intent/create";
-import { retrievePaymentIntent } from "./api/payment-intent/retrieve";
+import { createPaymentIntent } from "./api/intent/create";
+import { retrievePaymentIntent } from "./api/intent/retrieve";
 
 export const createPaymongoClient = (key: string) => {
   api.defaults.auth = {
@@ -13,7 +13,7 @@ export const createPaymongoClient = (key: string) => {
      * # PaymentIntent Resource
      * @link https://developers.paymongo.com/reference/the-payment-intent-object
      */
-    paymentIntent: {
+    intent: {
       create: createPaymentIntent,
       retrieve: retrievePaymentIntent,
     },
